Rename the tests
